api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
pytest==8.4.1
httpx==0.28.1
PyJWT==2.9.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
